Remove counter-productive unmatched asterisk test
The test was an edge case with unclear desired behavior.  As of now, I
can't think of a simple way to model the parser that would allow that
test to pass while still allowing a node to be nested within itself.

This is only an issue because both emphasis and stress use asterisks.