Pass failing blockquote test (wow)
In Outline/ParseBlockquote.ts, I was importing BlockQuoteNode.ts. In
Ast/BlockQuote.ts, I was importing BlockquoteNode.ts. Note the different
capitalizations of "q".

I suppose this caused two different versions of the class to be loaded
in memory, because the constructor from the Outline/ParseBlockQuote.ts
instantiation was not equal to the constructor from the
Ast/Blockquote.ts instantiation.

That's why there was an empty diff.