Move own anti-structural-typing fields to bottom
We add unique, unused fields to otherwise-identical syntax node classes
to differentiate them in the eyes of TypeScript's structural typing
system.

These fields have been moved to the bottom of each class.

We also add similar void methods to syntax node interfaces that
otherwise have no fields of their own (e.g.the InlineSyntaxNode
interface).

When a class implements one of those methods, it remains at the top of
the class, because it's essentially a declaration of the class's role.